Fix: Corrección errores en los permisos de los usuarios
Arreglados los permisos para los usuarios en la edicion y eliminación de documentos

+++ cocemfe_sevilla/documents/views.py
@@ -39,7 +39,7 @@ def upload_pdf(request):
                         messages.error(request, "La fecha de finalización debe ser posterior a la fecha actual.")
         else:
             form = PDFUploadForm()
-        return render(request, 'upload_pdf.html', {'form': form, 'superusers': professionals})
+        return render(request, 'upload_pdf.html', {'form': form, 'professionals_not_superuser': professionals})
     else:
         return render(request, '403.html')
 
@@ -62,27 +62,38 @@ def view_pdf_admin(request, pk):
 
 def update_pdf(request,pk):
     document = get_object_or_404(Document, pk=pk)
-    if request.method == 'POST':
-        form = PDFUploadForm(request.POST, instance=document)
-        if form.is_valid():
-            end_date = form.cleaned_data['end_date']
-            professionals = form.cleaned_data['professionals']
-            if end_date > timezone.now().date():
-                form.save()
-                document.professionals.set(professionals)
-                return redirect('list_pdf')
+    professionals_not_superuser = Professional.objects.filter(is_superuser=False)
+    if request.user.is_superuser:
+        if request.method == 'POST':
+            form = PDFUploadForm(request.POST, instance=document)
+            if form.is_valid():
+                end_date = form.cleaned_data['end_date']
+                professionals = form.cleaned_data['professionals']
+                for professional in professionals:
+                    if professional.is_superuser:
+                        messages.error(request, "Un administrador no puede ser seleccionado.")
+                if end_date > timezone.now().date():
+                    form.save()
+                    document.professionals.set(professionals)
+                    return redirect('list_pdf')
+                else:
+                    messages.error(request, "La fecha de finalización debe ser posterior a la fecha actual.")
             else:
-                messages.error(request, "La fecha de finalización debe ser posterior a la fecha actual.")
+                messages.error(request, "Por favor completa todos los campos del formulario.")
         else:
-            messages.error(request, "Por favor completa todos los campos del formulario.")
+            form = PDFUploadForm(instance=document)
+        return render(request, 'update_pdf.html', {'form': form, 'document': document, 'professionals_not_superuser': professionals_not_superuser})
     else:
-        form = PDFUploadForm(instance=document)
-    return render(request, 'update_pdf.html', {'form': form, 'document': document})
+        return render(request, '403.html')
 
 def delete_pdf(request, pk):
+    
     document = get_object_or_404(Document, pk=pk)
-    document.delete()
-    return redirect('list_pdf')   
+    if request.user.is_superuser:
+        document.delete()
+        return redirect('list_pdf')   
+    else:
+        return render(request, '403.html')
 
 def list_pdf(request):
     documentos = Document.objects.all()
